ci: update `labeler.yml` (#232)
This pull request includes a small change to the
`.github/workflows/labeler.yml` file. The change modifies the
`pull_request_target` event types to include `edited` and adds a
condition to the `Labeler` step to check for changes in the pull request
title.

*
[`.github/workflows/labeler.yml`](diffhunk://#diff-09b72f3c9a3e4f00ab00cd7000b302db25f056075d8895bd91b3654d6e7e956bL5-R13):
Updated `pull_request_target` event types to include `edited` and added
a condition to the `Labeler` step to check for changes in the pull
request title.