chore: update `dependabot.yml` configuration to include assignees, reviewers, and versioning strategy (#233)
This pull request includes changes to the `.github/dependabot.yml` file
to improve the management and organization of dependency updates. The
most important changes include the addition of assignees and reviewers,
the creation of dependency groups, and the specification of a versioning
strategy.

Improvements to dependency management:

* Added `lumirlumir` as an assignee for both npm and GitHub Actions
dependencies.
* Added `lumirlumir` as a reviewer for both npm and GitHub Actions
dependencies.
* Created dependency groups for `eslint-config-bananass`, `babel`,
`react`, and `next` to better organize and manage related dependencies.
* Added a versioning strategy set to 'increase' to ensure dependencies
are updated to the next version.